Show staking not rewarded in node description for node selection.
