+++ src/components/staking/StakingDialog.vue
@@ -84,7 +84,7 @@
                           class="h-is-text-size-1" style="border-radius: 4px"  @focus="stakeChoice='node'">
                   <option v-for="n in nodes" :key="n.node_id" :value="n.node_id"
                           style="background-color: var(--h-theme-box-background-color)">
-                    {{ makeNodeDescription(n) }} - {{ makeNodeStake(n) }}
+                    {{ makeNodeDescription(n) }} - {{ makeNodeStakeDescription(n) }}
                   </option>
                 </o-select>
               </o-field>
@@ -158,7 +158,7 @@
 import {computed, defineComponent, onMounted, PropType, ref, watch} from "vue";
 import {
   AccountBalanceTransactions,
-  AccountsResponse,
+  AccountsResponse, makeNodeStakeDescription,
   makeShortNodeDescription,
   NetworkNode
 } from "@/schemas/HederaSchemas";
@@ -299,27 +299,6 @@ export default defineComponent({
       return result
     }
 
-    const makeNodeStake = (node: NetworkNode) => {
-      const amountFormatter = new Intl.NumberFormat("en-US", {
-        maximumFractionDigits: 0
-      })
-      const percentFormatter = new Intl.NumberFormat("en-US", {
-        style: 'percent',
-        maximumFractionDigits: 1
-      })
-      const unclampedStakeAmount = ((node.stake_rewarded ?? 0) + (node.stake_not_rewarded ?? 0))/100000000
-      const percentMin = node.min_stake ? unclampedStakeAmount / (node.min_stake / 100000000) : 0
-      const percentMax = node.max_stake ? unclampedStakeAmount / (node.max_stake / 100000000) : 0
-
-      let result = amountFormatter.format(unclampedStakeAmount) + "ℏ staked"
-      if (percentMin != 0 && percentMin < 1) {
-        result += " (" + percentFormatter.format(percentMin) + " of Min)"
-      } else if (percentMax !== 0) {
-        result += " (" + percentFormatter.format(percentMax) + " of Max)"
-      }
-      return result
-    }
-
     const handleInput = (value: string) => {
       const previousValue = selectedAccount.value
       let isValidInput = true
@@ -410,7 +389,7 @@ export default defineComponent({
       handleCancelChange,
       handleConfirmChange,
       makeNodeDescription,
-      makeNodeStake,
+      makeNodeStakeDescription,
       handleInput
     }
   }
